requirements.txt
astroid==2.15.1
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
exceptiongroup==1.1.1
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
isort==5.11.5
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.33
packaging==23.1
platformdirs==3.2.0
pluggy==1.0.0
protobuf==3.20.3
pycodestyle==2.10.0
pylint==2.17.1
pytest==7.3.1
python-dotenv==0.21.1
requests==2.28.2
slack-bolt==1.17.0
slack-sdk==3.20.2
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
